Move json binder to `Response` class
object를 json으로 바인딩 하는 로직은 `Response` 클래스의 역할인 거 같아,
해당 로직을 `Response` class로 옮겼습니다.
